makw wallet a walletTab component
